requirements.txt
# --- Core Framework ---
Django==5.1.2
streamlit==1.40.0
python-dotenv==1.0.1
pandas==2.3.3
pandas-ta

# --- AI & Agents ---
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph
langgraph-checkpoint-redis
langchain-community>=0.3.0
google-generativeai>=0.8.0
tavily-python>=0.5.0
exa-py>=1.0.0
fastmcp>=0.1.0

# --- Databases & Infrastructure ---
redis>=5.0.1
redisvl>=0.3.0
psycopg2-binary>=2.9.9
pgvector>=0.3.6
timescale-vector>=0.0.1

# --- Utilities ---
httpx>=0.27.0
pydantic>=2.9.0
finnhub-python>=2.4.19
yfinance>=0.2.48